+++ tests/source_set_project/jsblock.js
@@ -1,0 +1,5 @@
+const { cube } = require('@jscad/modeling').primitives
+
+const main = () => cube({ size: 5 })
+
+module.exports = { main }